refactor: peer code suggestions
refactor dummy interface

#TISNEW-4109
