+++ src/app/dashboard/revalidation-history/PeriodicElement.ts
@@ -1,0 +1,10 @@
+export interface PeriodicElement {
+  recommendation: string;
+  episodeId: number;
+  outcome: string; // enum from GMC table
+  gmcSubDueDate: string;
+  ActSubDate: string;
+  submittedBy: string;
+  submissionStatus: string;
+  comments: string[];
+}
